Write VCPU_reg_get() and VCPU_reg_set()
